requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
Django==3.1.1
django-crispy-forms==1.9.2
django-fernet-fields==0.6
idna==2.10
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
twilio==6.45.4
urllib3==1.25.10
